A Scheme program to read logfiles
Just like at the dojo last week with Don, a program to read and parse logfiles for me.
So far I have a single file that will make sure that only a single argument is given.
That argument is supposed to be a filename. I can include this filename checker in the main file.
It'll be a good chance to play with including files, etc...
Next step is to remember how to read from files, should try and find twice.ss from the Winter.