Wow, ugly rebase. Too bad public: you get to see this This is a combination of 12 commits. prettier command line print
Used printf to output the argument followed by a linefeed.
Can't use (define ...) inside of the (if ...) statment, so I'll have to see how
it's handled in the file that includes this one.

working read from a file

cmd-args.scm will quit if there is an improper number of arguments.
logread.scm reads the first line from the filename given in the argument.

reading all lines in the file

Using a quick recursive function, I can read all lines from the file into a list.
Then I print that list for posterity. Now I should do some prettier formatting of the
line-strings and then print them in a prettier way.

delimited line input

Recursive function now delimits the syslog-formatted lines on input using the
"|" character, so that the date and the hostname can be seperated from the rest
of the message.

clearer docs for the logread parser

I've been thinking, and those are my thoughts. Next step should be reverting that silly
pipe insertion into the logfile line, and aquiring my tokenizer from CS241 in the Winter.

imported scanner from CS241

Also started DOT file to describe it, and made necessary little tweaks to
logread.scm to get rid of my earlier idea of delimiting with "|".

import of testeez framework, lots of cruft

There is lots of cruft because I added a bunch of transitions with the idea that I would
selectively tokenize dates and hostnames into tokens of their own type in the first pass.
But then wait, they are just strings, how can I tell a hostname from a word that's the
same as the hostname? So I'm going to first tokenize into words, then I can differentiate
later.

refactored and passed another test

working tokenize pass 2: first 3 tokens => 'date

A bunch of tests, and it works! Next I can add other functions to convert other fun
things into more ideal tokens. For example, I need to make sure that immediately
after the date token comes the hostname token. And then maybe daemon + PID tokens?

done converting the first 3 words into a single string

Had to massage it a little, but I inserted a space between all of the (listof char) in
the token, so that it can be list->string'd later, ready for use with the SRFI-19 date
struct stuff. Now, I can work on token-typing the rest of the line, and ... other
things?

intermediate state in upgrading scanner code for token-pass3

Need to get everything together in order to rebase this and commit to master.
Then I can push this to github.

uhh, file got deleted?